@@ -1,59 +1,59 @@
 {
-    "name": "ekindb-review",
-    "version": "0.8.1",
-    "license": "Apache-2.0",
-    "scripts": {
-        "ng": "ng",
-        "start": "ng serve",
-        "build": "ng build",
-        "test": "ng test",
-        "lint": "ng lint",
-        "e2e": "ng e2e"
-    },
-    "private": true,
-    "dependencies": {
-        "@angular/animations": "^17.0.5",
-        "@angular/cdk": "^17.0.2",
-        "@angular/common": "^17.0.5",
-        "@angular/compiler": "^17.0.5",
-        "@angular/core": "^17.0.5",
-        "@angular/forms": "^17.0.5",
-        "@angular/platform-browser": "^17.0.5",
-        "@angular/platform-browser-dynamic": "^17.0.5",
-        "@angular/router": "^17.0.5",
-        "@fullcalendar/angular": "^6.0.3",
-        "@fullcalendar/core": "^6.0.3",
-        "@fullcalendar/daygrid": "^6.0.3",
-        "@fullcalendar/interaction": "^6.0.3",
-        "@fullcalendar/timegrid": "^6.0.3",
-        "chart.js": "^3.3.2",
-        "file-saver": "^2.0.5",
-        "jspdf": "^2.5.1",
-        "jspdf-autotable": "^3.8.2",
-        "keycloak-angular": "^15.2.1",
-        "keycloak-js": "^22.0.5",
-        "primeflex": "^3.3.1",
-        "primeicons": "6.0.1",
-        "primeng": "17.2.0",
-        "prismjs": "^1.29.0",
-        "quill": "^1.3.7",
-        "rxjs": "~7.8.1",
-        "tslib": "^2.5.0",
-        "xlsx": "^0.18.5",
-        "zone.js": "~0.14.0"
-    },
-    "devDependencies": {
-        "@angular-devkit/build-angular": "^17.0.5",
-        "@angular/cli": "^17.0.5",
-        "@angular/compiler-cli": "^17.0.5",
-        "@types/file-saver": "^2.0.7",
-        "@types/jasmine": "~4.3.1",
-        "jasmine-core": "~4.6.0",
-        "karma": "~6.4.2",
-        "karma-chrome-launcher": "~3.2.0",
-        "karma-coverage": "~2.2.0",
-        "karma-jasmine": "~5.1.0",
-        "karma-jasmine-html-reporter": "~2.1.0",
-        "typescript": "~5.2.2"
-    }
+  "name": "ekindb-review",
+  "version": "0.8.1",
+  "license": "Apache-2.0",
+  "scripts": {
+    "ng": "ng",
+    "start": "ng serve",
+    "build": "ng build",
+    "test": "ng test",
+    "lint": "ng lint",
+    "e2e": "ng e2e"
+  },
+  "private": true,
+  "dependencies": {
+    "@angular/animations": "^18.0.2",
+    "@angular/cdk": "^18.0.2",
+    "@angular/common": "^18.0.2",
+    "@angular/compiler": "^18.0.2",
+    "@angular/core": "^18.0.2",
+    "@angular/forms": "^18.0.2",
+    "@angular/platform-browser": "^18.0.2",
+    "@angular/platform-browser-dynamic": "^18.0.2",
+    "@angular/router": "^18.0.2",
+    "@fullcalendar/angular": "^6.0.3",
+    "@fullcalendar/core": "^6.0.3",
+    "@fullcalendar/daygrid": "^6.0.3",
+    "@fullcalendar/interaction": "^6.0.3",
+    "@fullcalendar/timegrid": "^6.0.3",
+    "chart.js": "^3.3.2",
+    "file-saver": "^2.0.5",
+    "jspdf": "^2.5.1",
+    "jspdf-autotable": "^3.8.2",
+    "keycloak-angular": "^15.2.1",
+    "keycloak-js": "^22.0.5",
+    "primeflex": "^3.3.1",
+    "primeicons": "6.0.1",
+    "primeng": "17.2.0",
+    "prismjs": "^1.29.0",
+    "quill": "^1.3.7",
+    "rxjs": "~7.8.1",
+    "tslib": "^2.5.0",
+    "xlsx": "^0.18.5",
+    "zone.js": "~0.14.0"
+  },
+  "devDependencies": {
+    "@angular-devkit/build-angular": "^18.0.3",
+    "@angular/cli": "^18.0.3",
+    "@angular/compiler-cli": "^18.0.2",
+    "@types/file-saver": "^2.0.7",
+    "@types/jasmine": "~4.3.1",
+    "jasmine-core": "~4.6.0",
+    "karma": "~6.4.2",
+    "karma-chrome-launcher": "~3.2.0",
+    "karma-coverage": "~2.2.0",
+    "karma-jasmine": "~5.1.0",
+    "karma-jasmine-html-reporter": "~2.1.0",
+    "typescript": "~5.4.5"
+  }
 }
