@@ -1,5 +1,5 @@
 import {Injectable} from '@angular/core';
-import {HttpClient} from "@angular/common/http";
+import { HttpClient } from "@angular/common/http";
 import {Observable} from "rxjs";
 import {DatePipe} from "@angular/common";
 import {CountEntriesByDates} from "../api/countEntriesByDates";
